Setting up a way to talk with the Twitter API
Before the "meat" of the changes, I also standardized (will probably be
changed in the future) headers and footers of the pages.

Created the query page, that lets you search for specific words or
phrases in twitter. Needs a lot of work to sanitize input and add some
options like getting the retweets of that tweet or w/e.
By default it should just pull tweet text, display name and account name
of the author (I think, maybe this will change).
So far it doesnt get the usernames, simply because the standard query
doesnt pull them. Just changed the default query to pull them but need
to work on it.

The query page contacts with a middleware API made in Next,
/api/twitterquery (it will be executed serverside when its in
production, so user just sends a word to search for and gets the result
without any credential exposure (IN THEORY)) that handles contacting the
twitter API with the asked for parameters and fetching the response
data.

Need to create an .env.local (.env should work I think) with the dev
BEARER_TOKEN

And that's it. Next step is adding some decent css to the page, matching
usernames with tweets and then checking what the objectives are. Then
maybe hooking the db up or w/e. Long way. Yay