Dynamic routing tweets, saving searched tweets
Lets you click on any tweet and it'll load it up in a page (need to get
more tweet info from the api, figure out how their imgs are routed to
show profile pages).
Can show a lot of info there, need to think what we need from API.

Tweets are now saved in sessionStorage (doesnt expire until tab or
browser is closed). I'm using react contexts for this.

Structured everything a bit more clearly, now the search and old tweets
main pages go in their folders as index.js. In the case of the search,
tweets are routed via the dynamic [id].js route, checking first if their
routed ids are saved in the state.

Added a file of aux functions for tweet data handling. It probably
should not go into components but in a folder called lib or something.

Improved a bit the css and we're now using the "json" file of dummydata
(it should not be prettyfied json and also it should just be a json
file, not a js file, but convenience w/e).