Show chosen tweets, use "big" query, misc improvements
It now shows clicked on tweets after searching. It shows them using the
embed api (it gives you the tweet html and then you populate its attrs
via an external script) though I had to do weird stuff because this is
react, and it hates when the DOM is changed without his permission.
Also added a extra tweet info box but its not working yet.

We communicate with the twitter api via another middlepoint api,
getEmbeddedTweet. Good thing is, it doesnt require auth. So we get the
real text of a tweet without having to use the query with everything.

The megaquery allows us to get more info on tweets, like date and stuff,
which is also shown.

Data is now transferred through pages via context inside _app (in normal
react it'd be a bit different, but it seems its the way to do it in
next)

Changed name of a lot of stuff to make everything clearer. Removed the
placeholder tweet stuff, since we're using the "real" tweet style now.

Probably something else I'm forgetting.