Did you know...
That linters will complain if your react class doesnt start with a
majuscule? Wooo this was a fun one.
Also moved a bit the logic of a useEffect so it always triggers the
logic inside the useEffect instead of checking for stuff outside of it.
Its ready to be uploaded to vercel.