Fixed main page divs, added send msg to telegram functionality
Note that the button and stuff to send msgs to telegram will be moved to
another place and made pretty. I also added another env variable to .env
to make it work. the variable is chatID, which should be an input with a
default.
So except for the changes to the redirect divs everything is temporal,
just want to see it work.